Don't allow installing/updating/removing extension mode
Prevent updating mode extensions installing in user's directory and report
error accordingly (dbus and direct shell calls). Users can still go on
another session (where this extension isn't a mode one) to be prompted.
However, when switching back to session having this extension as a mode,
the system one will be loaded. Same for mechanism for uninstalling or updating.