VolumeOverrideDialog to enable volume above 100%
This dialog is triggered by a dbus call and just change directly the gsettings
amplified key if acked by the user.
There is as well some description and link to the sound settings in Settings
for people to check their global volume.

The dbus call will be emitted by gnome-settings-daemon for instance if the
volume plus media key is triggered multiple consecutive times at maximum volume.